requirements.txt
flask
openai
python-dotenv
azure-search-documents